Fix the slack link
There are two slack links in the community page, both point to the
sig-events channel on the CDF slack. Those links do not work for those
who have not joined the CDF slack yet.

Replace the links with a link to the public invite to the slack
workspace. Add a comment about the #cdevents channel, which is where
CDEvents discussions happen today.

Fixes: #40

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>